Revert UserGroupForm parent class to FormType
UserGroup entity does not have any language support currently
implemented. EntityFormType class presumes that the current Entity is
translatable, but when trying to create a new group will result in an
error.
The easiest non-rippling fix here is to revert to the older FormType
parent class. This might be reverted in the future if UserGroup is made
translatable or EntityFormType will have support for non-translatable
entities.