Move setDefaultLangcode method to LibraryTrait (#38)
Adding a new service point was failing with a SQL NOT NULL constraint
violation because the associated addresses were not inheriting the
default language code from the parent ServicePoint entity. Both Library
and ServicePoint entities rely on LibraryTrait for the LibraryInterface
implementation but only Library had a default language code setter
method that took care of setting the default language code also for the
addresses. Move this setter method to be a part of LibraryTrait to make
it available also for ServicePoint.