Add library system info fields to Library (#36)
These fields record the library system and a system-specific identifier
for a library or a service point. This is used for combining data from
library directory and other APIs for statistical purposes. At first this
will be done with Axiell's API.

Note: The required filtering functionality for Library Directory API
is implemented in API v4. (libraries-fi/kirkanta-api#10)